requirements.txt
streamlit>=1.30.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
google-search-results>=2.4.2
imagehash>=4.3.1
Pillow>=10.0.0
requests>=2.31.0
pandas>=2.0.0
